+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Change log for xITGHyperV
 
-## Unreleased
+## 1.1
+
+New features
+
++ xVMComPort DSC Resource
+
+## 1.0
+
+New features
+
++ xVMLegacyNetworkAdapter DSC Resource